Fixes up the adding connections view. Fixes #5.
Somebody though it was a good idea to create an entire Page when they couldn't find
the object associated with a particular vanity url. I don't get how that made sense
to someone at the time, but I've ripped that out. If it can't find either of the
models its trying to connect then it bails as it should.